+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - More modules: `memoizeStringOnly`, `joinClasses`
+- `UserAgent`: Query information about current user agent
 
 ### Changed
 - `fetchWithRetries`: Reject failure with an Error, not the response